pyproject.toml
[tool.poetry]
name = "api.hunj.tech"
version = "0.1.0"
description = "personal api"
authors = ["hunj <hunjaelee@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.4"
pillow = "^10.2.0"
psycopg2 = "^2.9.7"
gunicorn = "^21.2.0"
djangorestframework = "^3.14.0"
celery = {extras = ["redis"], version = "^5.3.5"}
redis = "^5.0.1"
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
